"refactor: reorganiza arquivos da View"
